requirements.txt
amqp==2.4.2
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==2.1.7
django-debug-toolbar==1.11
django-redis==4.10.0
django-rest-swagger==2.2.0
djangorestframework==3.9.2
docopt==0.6.2
et-xmlfile==1.0.1
idna==2.8
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
kombu==4.4.0
MarkupSafe==1.1.1
odfpy==1.4.0
openapi-codec==1.3.2
openpyxl==2.4.11
pkg-resources==0.0.0
Pygments==2.3.1
pytz==2018.9
PyYAML==3.13
records==0.5.3
redis==3.2.0
requests==2.21.0
simplejson==3.16.0
SQLAlchemy==1.3.0
sqlparse==0.2.4
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.24.1
vine==1.2.0
xlrd==1.2.0
xlwt==1.3.0
